refactor cast
